+++ CHANGELOG.md
@@ -14,6 +14,9 @@ Docker images:
 * Upgrade Python used by Docker images from 3.8.10 to 3.8.12.
 * Base distribution version for non slim Alpine Linux based images has been upgraded from Debian Buster (10) to Debian Bullseye.
 
+Other:
+* Support for ``ujson`` JSON library (``json_library`` configuration option) has been removed in favor of ``orjson``.
+
 ## 2.1.26 "Yavin" - Jan 12, 2022
 
 <!---